Added news pagination and template tweaks
