+++ remedy/admin_views/newsview.py
@@ -1,22 +1,13 @@
 from admin_helpers import *
 
 from flask.ext.admin.contrib.sqla import ModelView
-
 from remedy.rad.models import News
-from wtforms import fields, widgets
-
-
-class CKTextAreaWidget(widgets.TextArea):
-    def __call__(self, field, **kwargs):
-        kwargs['class'] = 'form-control ckeditor'
-        return super(CKTextAreaWidget, self).__call__(field, **kwargs)
-
-
-class CKTextAreaField(fields.TextAreaField):
-    widget = CKTextAreaWidget()
+from remedy.rad.cktextarea import CKTextAreaField
 
 
 class NewsView(AdminAuthMixin, ModelView):
+
+    can_view_details = True
 
     form_overrides = dict(text=CKTextAreaField)
 
@@ -24,10 +15,10 @@ class NewsView(AdminAuthMixin, ModelView):
     edit_template = 'admin/news_edit.html'
 
     form_excluded_columns = (
-        'date_created', 'visible'
+        'date_created'
     )
 
-    column_exclude_list = ('visible')
+    column_exclude_list = ()
 
     def __init__(self, session, **kwargs):
         super(NewsView, self).__init__(News, session, **kwargs)
