+++ remedy/rad/cktextarea.py
@@ -1,0 +1,12 @@
+
+from wtforms import fields, widgets
+
+
+class CKTextAreaWidget(widgets.TextArea):
+    def __call__(self, field, **kwargs):
+        kwargs['class'] = 'form-control ckeditor'
+        return super(CKTextAreaWidget, self).__call__(field, **kwargs)
+
+
+class CKTextAreaField(fields.TextAreaField):
+    widget = CKTextAreaWidget()
