+++ remedy/templates/admin/news_edit.html
@@ -2,5 +2,5 @@
 
 {% block tail %}
     {{ super() }}
-    <script src="http://cdnjs.cloudflare.com/ajax/libs/ckeditor/4.0.1/ckeditor.js"></script>
+    <script src="http://cdnjs.cloudflare.com/ajax/libs/ckeditor/4.5.4/ckeditor.js"></script>
 {% endblock %}
