+++ remedy/templates/base.html
@@ -134,14 +134,14 @@
           #}
           <li role="presentation" class="dropdown">
 		  
-		  <li><a href="{{ url_for('remedy.news') }}">News</a></li>
-		  
             <a class="dropdown-toggle" data-toggle="dropdown" href="#" role="button" aria-expanded="false">About <span class="caret"></span></a>
             <ul class="dropdown-menu" role="menu">
               <li><a href="{{ url_for('remedy.about') }}">About RAD Remedy</a></li>
               <li><a href="{{ url_for('remedy.about_the_beta') }}">About the RAD Beta Launch</a></li>
+			  <li><a href="{{ url_for('remedy.projects') }}">RAD Projects</a></li>
             </ul>
-          <li><a href="{{ url_for('remedy.projects') }}">RAD Projects</a></li>
+			
+			<li><a href="{{ url_for('remedy.news') }}">News</a></li>
           <li><a href="{{ url_for('remedy.get_involved') }}">Get Involved</a></li>
           <li><a href="{{ url_for('remedy.resource_search', autofill='1') }}">Find Resources <em>(Beta!)</em></a></li>
           <li><a href="{{ url_for('remedy.submit_provider') }}">Submit a Provider</a></li>
@@ -159,6 +159,7 @@
               <li><a href="{{ url_for('remedy.how_to_use') }}">How to Use RAD</a></li>
               #}
               <li><a href="{{ url_for('remedy.about') }}">About RAD Remedy</a></li>
+			  <li><a href="{{ url_for('remedy.news') }}">News</a></li>
               <li><a href="{{ url_for('remedy.about_the_beta') }}">About the RAD Beta Launch</a></li>
               <li><a href="{{ url_for('remedy.projects') }}">RAD Projects</a></li>
               <li><a href="{{ url_for('remedy.get_involved') }}">Get Involved</a></li>
