+++ remedy/templates/find-provider.html
@@ -168,11 +168,6 @@
         {{ macros.render_pagination(pagination) }}
       </div>
 
-	<div class="col-md-6">
-This map only shows results within this page
-</div>
-
-
       <div class="col-md-6">
         <div id="provider-map"></div>
       </div>
