+++ remedy/templates/news-item.html
@@ -3,27 +3,18 @@
 
 {% block content %}
 
-<h2>Vestibulum ornare suscipit tempu</h2>
-
+<h2>{{ news.subject }}</h2>
 
 
 <p>
-<small>Posted on September 3, 2016 by <strong>raduser</strong></small>
+<small>Posted on {{ news.date_created.strftime('%m/%d/%Y') }} by <strong>{{ news.author }}</strong></small>
 </p>
-
 
 <hr>
 
-
 <div>
-
-
-              <p>Donec consectetur accumsan ipsum, eu luctus augue aliquet vitae. Donec mollis venenatis libero, et venenatis arcu semper eget. Phasellus metus est, feugiat ut vulputate a, varius eu nulla. Proin eu pharetra nibh. Quisque vehicula dolor est, non commodo ligula. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum volutpat rhoncus quam quis vehicula. In fermentum venenatis lacus, sed viverra metus accumsan id. Nulla aliquet, ante sed interdum fermentum, orci nulla bibendum sapien, a vehicula nulla turpis vel leo. Duis dapibus eleifend posuere. Sed sem orci, dignissim eu vehicula ut, suscipit at augue. Maecenas sed sapien at quam sodales porttitor non ac sapien. Maecenas et dui eros, non consectetur dui. Mauris laoreet lacinia consequat. Aliquam posuere interdum ante, vel vehicula dui sodales a. Donec in leo quis neque imperdiet volutpat vitae quis lacus.</p>
-              <p>Suspendisse id urna dui, ut vestibulum massa. Sed et nibh quis tellus bibendum pharetra ut ac nisl. Nulla malesuada tellus euismod eros faucibus gravida. Aliquam eu nulla sit amet ligula sagittis pretium. Mauris vitae purus nisi. Aenean quam massa, convallis vitae fringilla sed, aliquam non lectus. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Suspendisse potenti. Duis tincidunt cursus elit, vel faucibus lectus feugiat a. Integer pretium dui ut ligula bibendum suscipit. Vestibulum ornare suscipit tempus. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.</p>
-              <p>In ipsum lectus, pharetra vel rutrum at, gravida sollicitudin ipsum. Nam id pharetra felis. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed feugiat ligula id augue gravida dapibus. Quisque odio lectus, euismod et malesuada et, ultrices ac libero. Quisque elit lacus, fringilla quis adipiscing vitae, laoreet eu neque. Vestibulum bibendum, turpis vestibulum hendrerit ultrices, justo est semper sapien, eget fermentum enim arcu nec eros. Fusce lobortis velit eros. Ut non eros eget est gravida posuere.</p>
-			  
-			  
-
+{{ news.text|safe }}
 </div>
+
 
 {% endblock %}
